+++ packages/config-plugins/src/ios/__tests__/Google-test.ts
@@ -1,10 +1,8 @@
 import {
   getGoogleMapsApiKey,
-  getGoogleMobileAdsAppId,
   getGoogleServicesFile,
   getGoogleSignInReservedClientId,
   setGoogleMapsApiKey,
-  setGoogleMobileAdsAppId,
   setGoogleSignInReservedClientId,
 } from '../Google';
 import { appendScheme } from '../Scheme';
@@ -14,16 +12,13 @@ jest.mock('../Scheme');
 describe('ios google config', () => {
   it(`returns null from all getters if no value provided`, () => {
     expect(getGoogleMapsApiKey({})).toBe(null);
-    expect(getGoogleMobileAdsAppId({})).toBe(null);
     expect(getGoogleSignInReservedClientId({})).toBe(null);
     expect(getGoogleServicesFile({})).toBe(null);
   });
 
   it(`returns the correct values from all getters if a value is provided`, () => {
     expect(getGoogleMapsApiKey({ ios: { config: { googleMapsApiKey: '123' } } })).toBe('123');
-    expect(getGoogleMobileAdsAppId({ ios: { config: { googleMobileAdsAppId: 'abc' } } })).toBe(
-      'abc'
-    );
+
     expect(
       getGoogleSignInReservedClientId({
         ios: { config: { googleSignIn: { reservedClientId: '000' } } },
@@ -42,16 +37,6 @@ describe('ios google config', () => {
     );
 
     expect(setGoogleMapsApiKey({}, {})).toMatchObject({});
-  });
-
-  it(`sets the google mobile ads app id if provided or returns plist`, () => {
-    expect(
-      setGoogleMobileAdsAppId({ ios: { config: { googleMobileAdsAppId: '123' } } }, {})
-    ).toMatchObject({
-      GADApplicationIdentifier: '123',
-    });
-
-    expect(setGoogleMobileAdsAppId({}, {})).toMatchObject({});
   });
 
   it(`adds the reserved client id to scheme if provided`, () => {
