+++ packages/config-plugins/src/plugins/unversioned/expo-ads-admob.ts
@@ -1,0 +1,23 @@
+import { ConfigPlugin } from '../../Plugin.types';
+import { withAdMob as withAdMobAndroid } from '../../android/AdMob';
+import { withAdMob as withAdMobIOS } from '../../ios/AdMob';
+import { createRunOncePlugin } from '../core-plugins';
+import { withStaticPlugin } from '../static-plugins';
+
+const packageName = 'expo-ads-admob';
+
+export const withAdMob: ConfigPlugin = config => {
+  return withStaticPlugin(config, {
+    plugin: packageName,
+    // If the static plugin isn't found, use the unversioned one.
+    fallback: withUnversionedAdMob,
+  });
+};
+
+const withUnversionedAdMob: ConfigPlugin = createRunOncePlugin(config => {
+  config = withAdMobAndroid(config);
+  config = withAdMobIOS(config);
+  return config;
+}, packageName);
+
+export default withAdMob;
