+++ packages/expo-cli/src/commands/apply/configureProjectAsync.ts
@@ -29,7 +29,7 @@ export const expoManagedPlugins = [
   'expo-file-system',
   'expo-location',
   'expo-media-library',
-  'expo-notifications',
+  // 'expo-notifications',
   'expo-screen-orientation',
   'expo-sensors',
   'expo-task-manager',
@@ -43,7 +43,7 @@ const expoManagedVersionedPlugins = [
   // 'expo-facebook',
   // 'expo-branch',
   // 'expo-updates',
-  'expo-ads-admob',
+  // 'expo-ads-admob',
   'expo-apple-authentication',
   'expo-document-picker',
   'expo-firebase-analytics',
